fix: only allow deleting by note owner

+++ notes-backend/controllers/notes.js
@@ -48,7 +48,26 @@ notesRouter.post("/", async (request, response) => {
 });
 
 notesRouter.delete("/:id", async (request, response) => {
+  // only allow deletion if user is the owner of the note
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  // check if user is the owner of the note
+  const user = await User.findById(decodedToken.id);
+  const note = await Note.findById(request.params.id);
+
+  // if not, return 403
+  if (user.id !== note.user.toString()) {
+    return response.status(403).json({ error: "permission denied" });
+  }
+
+  // if so, delete
   await Note.findByIdAndDelete(request.params.id);
+  // remove note from user
+  user.notes = user.notes.filter((note) => note.id !== request.params.id);
+  await user.save();
   response.status(204).end();
 });
